Move Emacs installation locations into canonical places.
This means that run_emacs.py has less work finding them.
